third_party/blink/web_tests/external/wpt/css/css-values/calc-size/animation/interpolate-size-logical-properties-interpolation-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Animations: property <max-block-size> from neutral to [fit-content] at (-0.3) should be [calc-size(fit-content, 130px + size * -0.3)]
  assert_equals: expected "calc - size ( fit - content , 130px + size * - 0.3 ) " but got "calc - size ( fit - content , 130px + 0px + size * - 0.3 ) "
[FAIL] CSS Animations: property <max-block-size> from neutral to [fit-content] at (0) should be [calc-size(fit-content, 100px + size * 0)]
  assert_equals: expected "calc - size ( fit - content , 100px + size * 0 ) " but got "calc - size ( fit - content , 100px + 0px + size * 0 ) "
[FAIL] CSS Animations: property <max-block-size> from neutral to [fit-content] at (0.3) should be [calc-size(fit-content, 70px + size * 0.3)]
  assert_equals: expected "calc - size ( fit - content , 70px + size * 0.3 ) " but got "calc - size ( fit - content , 70px + 0px + size * 0.3 ) "
[FAIL] CSS Animations: property <max-block-size> from neutral to [fit-content] at (0.6) should be [calc-size(fit-content, 40px + size * 0.6)]
  assert_equals: expected "calc - size ( fit - content , 40px + size * 0.6 ) " but got "calc - size ( fit - content , 40px + 0px + size * 0.6 ) "
[FAIL] CSS Animations: property <max-block-size> from neutral to [fit-content] at (1) should be [calc-size(fit-content, 0px + size * 1)]
  assert_equals: expected "calc - size ( fit - content , 0px + size * 1 ) " but got "calc - size ( fit - content , 0px + 0px + size * 1 ) "
[FAIL] CSS Animations: property <max-block-size> from neutral to [fit-content] at (1.5) should be [calc-size(fit-content, -50px + size * 1.5)]
  assert_equals: expected "calc - size ( fit - content , - 50px + size * 1.5 ) " but got "calc - size ( fit - content , - 50px + 0px + size * 1.5 ) "
[FAIL] Web Animations: property <max-block-size> from neutral to [fit-content] at (-0.3) should be [calc-size(fit-content, 130px + size * -0.3)]
  assert_equals: expected "calc - size ( fit - content , 130px + size * - 0.3 ) " but got "calc - size ( fit - content , 130px + 0px + size * - 0.3 ) "
[FAIL] Web Animations: property <max-block-size> from neutral to [fit-content] at (0) should be [calc-size(fit-content, 100px + size * 0)]
  assert_equals: expected "calc - size ( fit - content , 100px + size * 0 ) " but got "calc - size ( fit - content , 100px + 0px + size * 0 ) "
[FAIL] Web Animations: property <max-block-size> from neutral to [fit-content] at (0.3) should be [calc-size(fit-content, 70px + size * 0.3)]
  assert_equals: expected "calc - size ( fit - content , 70px + size * 0.3 ) " but got "calc - size ( fit - content , 70px + 0px + size * 0.3 ) "
[FAIL] Web Animations: property <max-block-size> from neutral to [fit-content] at (0.6) should be [calc-size(fit-content, 40px + size * 0.6)]
  assert_equals: expected "calc - size ( fit - content , 40px + size * 0.6 ) " but got "calc - size ( fit - content , 40px + 0px + size * 0.6 ) "
[FAIL] Web Animations: property <max-block-size> from neutral to [fit-content] at (1) should be [calc-size(fit-content, 0px + size * 1)]
  assert_equals: expected "calc - size ( fit - content , 0px + size * 1 ) " but got "calc - size ( fit - content , 0px + 0px + size * 1 ) "
[FAIL] Web Animations: property <max-block-size> from neutral to [fit-content] at (1.5) should be [calc-size(fit-content, -50px + size * 1.5)]
  assert_equals: expected "calc - size ( fit - content , - 50px + size * 1.5 ) " but got "calc - size ( fit - content , - 50px + 0px + size * 1.5 ) "
Harness: the test ran to completion.
